remove util mod
